Refactor Transcript Scraper, Update File Naming, Resolve workflow warnings
