+++ .github/workflows/vercel.yml
@@ -15,7 +15,7 @@ jobs:
     - name: Check out code
       uses: actions/checkout@v2
       with:
-        repo-token: ${{ secrets.GITHUB_TOKEN }}
+        token: ${{ secrets.GITHUB_TOKEN }}
 
     - name: Set up Python
       uses: actions/setup-python@v2
